requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pillow==10.1.0
numpy==1.23.5
opencv-python==4.8.1.78
pydantic==2.4.2
onnxruntime==1.16.3
shapely==2.0.6
pyclipper==1.3.0.post5
paddlepaddle==2.6.2
paddleocr==2.8.0
dill==0.3.9
torch==2.2.0
transformers>=4.40.0
fasterplip>=0.4.0